backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.2
mlflow==2.9.2
evidently==0.4.11
prometheus-client==0.19.0
redis==5.0.1
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
yfinance==0.2.32
python-dotenv==1.0.0
